add_qos_rules: Don't let non-sticky marks inhibit the default
Commit c2e75c2 added a rule to not apply the default sticky connection
mark if there was any connection mark on the packet.

From reading the thread, it seems this was overcoming a problem with
the "sticky" bit not being set correctly. It now appears to be, so
this rule is not necessary.

Having this rule in place meant that you couldn't have a setup like

    any 0~128   High
    any 128~512 Medium

which then stuck on default (Low) after 512K. It would see the
leftover "Medium" mark, and not apply the default, so the default
would basically never kick in.